+++ db/dB_temp/planifications.csv
@@ -3,3 +3,6 @@ id,date,user_id,status,routine_name
 2,2024-06-02,2,True,Cardio
 3,2024-06-03,1,False,Strength
 4,2024-05-20,2,False,Pecho espalda
+5,2024-08-25,1,False,Pecho
+6,2024-08-10,2,False,JuanitoAlimaña
+7,2024-08-23,1,False,Piernas
